Use "core" format for home page
The first page of MHCLG Way currently duplicates top level headings in the left-hand navigation.

We want to suppress the auto-generated contents list for this first page.  The default layout ("layout") includes this, but the "core" layout makes the sidebar configurable with no default content.

See the tech docs documentation site for details.  The site is sadly no longer deployed anywhere, but guidance still available on GitHub :(
https://github.com/alphagov/tdt-documentation/blob/main/source/configure_project/frontmatter/index.html.md.erb#L107-L119