Review and update Git and GitHub pages
I've added some sections to the working in GitHub page - the "transferring in" advice comes from a recent lesson learned the hard way when someone transferred in about 50 repositories and lost access to them all because he wasn't listed explicitly in the repo users!

Have updated the review dates on all pages as these seem good for now.