Update python recommendation to use `uv`
`uv` is the latest tool for managing Python projects. It replaces a
swathe of tools that otherwise need to be managed together: pip, pyenv,
virtualenvs, pip-tools, etc.

While a relatively new entry, it has extremely good support and
community uptake, and has been used by teams in MHCLG for 6 months+ with
no concerns and lots of upsides.

I believe this should be the general recommendation going forwards as it
will provide:

- consistency between Python teams in the department
- a better developer experience vs managing multiple tools
- more streamlined dependency management